LayoutTests/crypto/subtle/aes-cbc-encrypt-decrypt.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test AES-CBC encrypt and decrypt functions.");

jsTestIsAsync = true;

if (!window.subtle)
    window.crypto.subtle = window.crypto.webkitSubtle;

var keyData = hexToArrayBuffer("2b7e151628aed2a6abf7158809cf4f3c");
var iv = hexToArrayBuffer("000102030405060708090a0b0c0d0e0f");
var plaintext = hexToArrayBuffer("6bc1bee22e409f96e93d7e117393172aae2d8a571e03ac9c9eb76fac45af8e5130c81c46a35ce411e5fbc1191a0a52eff69f2445df4f9b17ad2b417be66c3710");
var extractable = true;

debug("Importing a raw AES key from string literal...");
crypto.subtle.importKey("raw", keyData, "aes-cbc", extractable, ["encrypt", "decrypt"]).then(function(result) {
    key = result;
    shouldBe("key.type", "'secret'");
    shouldBe("key.extractable", "true");
    shouldBe("key.algorithm.name", "'aes-cbc'");
    shouldBe("key.algorithm.length", "128");
    shouldBe("key.usages", "['encrypt', 'decrypt']");

    debug("Using the key to encrypt plaintext...");
    return crypto.subtle.encrypt({name: "aes-cbc", iv: iv}, key, [plaintext]);
}).then(function(result) {
    encryptionResult = result;
    shouldBe("byteArrayToHexString(new Uint8Array(encryptionResult))", "'[76 49 ab ac 81 19 b2 46 ce e9 8e 9b 12 e9 19 7d 50 86 cb 9b 50 72 19 ee 95 db 11 3a 91 76 78 b2 73 be d6 b8 e3 c1 74 3b 71 16 e6 9e 22 22 95 16 3f f1 ca a1 68 1f ac 09 12 0e ca 30 75 86 e1 a7 8c b8 28 07 23 0e 13 21 d3 fa e0 0d 18 cc 20 12]'");

    debug("Decrypting it back...");
    return crypto.subtle.decrypt({name: "aes-cbc", iv: iv}, key, [result]);
}).then(function(result) {
    decryptionResult = result;
    shouldBe("new Uint8Array(decryptionResult)", "plaintext");

    debug("Testing initialization vector bindings...");
    shouldThrow("crypto.subtle.encrypt({name: 'AES-CBC', iv: null}, key, [plaintext])");
    shouldThrow("crypto.subtle.encrypt({name: 'AES-CBC'}, key, [plaintext])");
    shouldThrow("crypto.subtle.encrypt({name: 'AES-CBC', iv: 3}, key, [plaintext])");
    shouldThrow("crypto.subtle.encrypt({name: 'AES-CBC', iv: new Uint8Array([0])}, key, [plaintext])");

    finishJSTest();
});
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
